+++ api/src/setup.rs
@@ -1,14 +1,12 @@
 // The Public Setup needed for Proofs
 use crate::anon_xfr::{
-    abar_to_bar::build_abar_to_bar_cs,
     abar_to_ar::build_abar_to_ar_cs,
-    circuits::{
-        build_multi_xfr_cs, AMultiXfrWitness, PayeeSecret, PayerSecret, TurboPlonkCS, TREE_DEPTH,
-    },
-    bar_to_abar::build_bar_to_abar_cs,
-    config::FEE_TYPE,
-    structs::{MTNode, MTPath},
+    abar_to_bar::build_abar_to_bar_cs,
     ar_to_abar::build_ar_to_abar_cs,
+    bar_to_abar::build_bar_to_abar_cs,
+    circuits::build_multi_xfr_cs,
+    structs::{MTNode, MTPath},
+    PayeeSecret, PayerSecret, TurboPlonkCS, FEE_TYPE, TREE_DEPTH,
 };
 use crate::parameters::{
     ABAR_TO_AR_VERIFIER_PARAMS, ABAR_TO_BAR_VERIFIER_PARAMS, AR_TO_ABAR_VERIFIER_PARAMS,
@@ -31,6 +29,8 @@ use zei_plonk::{
     },
     poly_commit::{kzg_poly_com::KZGCommitmentSchemeBLS, pcs::PolyComScheme},
 };
+// The Public Setup needed for Proofs
+use crate::anon_xfr::abar_to_abar::AMultiXfrWitness;
 
 // Shared by all members of the ledger
 #[derive(Serialize, Deserialize)]
@@ -448,7 +448,7 @@ impl From<ProverParams> for VerifierParams {
 
 #[cfg(test)]
 mod test {
-    use crate::anon_xfr::circuits::TREE_DEPTH;
+    use crate::anon_xfr::TREE_DEPTH;
     use crate::parameters::SRS;
     use crate::setup::{ProverParams, VerifierParams};
     use zei_algebra::{
